+++ App_Function_Libraries/Benchmarks_Evaluations/Confabulation_check.py
@@ -14,7 +14,7 @@
 #
 # Functions:
 from App_Function_Libraries.Chat import chat_api_call
-from App_Function_Libraries.ms_g_eval import validate_inputs, detailed_api_error
+from App_Function_Libraries.Benchmarks_Evaluations.ms_g_eval import validate_inputs, detailed_api_error
 
 
 def simplified_geval(transcript: str, summary: str, api_name: str, api_key: str, temp: float = 0.7) -> str:
